+++ tests/conftest.py
@@ -1,4 +1,6 @@
 import clickhouse_connect
+import numpy as np
+import pandas as pd
 import splink.comparison_library as cl
 from chdb import dbapi
 from pytest import fixture, mark, param
@@ -7,6 +9,22 @@ from splink import ColumnExpression, SettingsCreator, block_on, splink_datasets
 from splinkclickhouse import ChDBAPI, ClickhouseAPI
 
 df = splink_datasets.fake_1000
+
+np.random.seed(2542546873)
+
+
+def pytest_collection_modifyitems(items, config):
+    # anything marked with chdb will also have chdb_only, and vice versa
+    # so don't worry about those, and then they don't get added to core tests
+    our_marks = {"chdb", "clickhouse"}
+
+    for item in items:
+        # any test without our marks is core.
+        # Runs on e.g. -m chdb by not on -m chdb_no_core
+        if not any(marker.name in our_marks for marker in item.iter_markers()):
+            item.add_marker("core")
+            for mark in our_marks:
+                item.add_marker(mark)
 
 
 @fixture
@@ -45,8 +63,8 @@ def clickhouse_api():
 
 @fixture(
     params=[
-        param("chdb", marks=[mark.chdb]),
-        param("clickhouse", marks=[mark.clickhouse]),
+        param("chdb", marks=[mark.chdb, mark.chdb_no_core]),
+        param("clickhouse", marks=[mark.clickhouse, mark.clickhouse_no_core]),
     ]
 )
 def api_info(request, chdb_api, clickhouse_api):
@@ -112,3 +130,26 @@ def fake_1000_settings_factory():
         )
 
     return fake_1000_settings
+
+
+@fixture
+def input_nodes_with_lat_longs():
+    lat_low, lat_high = 49, 61
+    long_low, long_high = -8, 2
+
+    n_rows = 1_000
+    lats = np.random.uniform(low=lat_low, high=lat_high, size=n_rows)
+    longs = np.random.uniform(low=long_low, high=long_high, size=n_rows)
+    # also include some names so we have a second comparison
+    names = np.random.choice(
+        ("tom", "tim", "jen", "jan", "ken", "sam", "katherine"),
+        size=n_rows,
+    )
+    return pd.DataFrame(
+        {
+            "unique_id": range(n_rows),
+            "name": names,
+            "latitude": lats,
+            "longitude": longs,
+        }
+    )